Move verifyTempaltes to it's own file, semantic naming
